补充下单的操作

+++ ddd/src/main/java/com/ddd/user/infrastructure/mapper/UserMapper.java
@@ -19,6 +19,6 @@ public interface UserMapper {
      * @return 用户数据
      */
     @Select("select * from user where id = #{id}")
-    UserDO selectUserById(Integer id);
+    UserDO selectUserById(Long id);
 
 }
